// app/Engine/Types.ts
export { Type };

enum Type
{
    Game = "Game",
    Scene = "Scene",
    Scene2D = "Scene2D",
    SceneObject = "SceneObject",
    DrawObject = "DrawObject",
    Line = "Line",
    ImageObject = "ImageObject",
    Sprite = "Sprite",
    Tile = "Tile",
    Light = "Light",
    SpotLight = "SpotLight",
    PointLight = "PointLight",
    DirectionalLight = "DirectionalLight",
    SoundObject = "SoundObject",
    UIControlObject = "UIControlObject",
    Material = "Material",
    NodeMaterial = "NodeMaterial",
    PlanarMaterial = "PlanarMaterial",
    ShaderMaterial = "ShaderMaterial",
    MaterialNode = "MaterialNode",
    ImageCollection = "ImageCollection",
    SpriteSet = "SpriteSet",
    SpriteSetCollection = "SpriteSetCollection"
}